+++ Kernel/System/DynamicField/Driver/Checkbox.pm
@@ -297,12 +297,15 @@ EOF
         my $TranslatedDesc = $Param{LayoutObject}->{LanguageObject}->Translate(
             'Ignore this field.',
         );
-        $HTMLString = <<"EOF";
+
+        if ( !$Param{NoIgnoreField} ) {
+            $HTMLString = <<"EOF";
 <input type="radio" id="$FieldNameUsed0" name="$FieldNameUsed" value="" $FieldUsedChecked0 />
 $TranslatedDesc
 <div class="clear"></div>
 <input type="radio" id="$FieldNameUsed1" name="$FieldNameUsed" value="1" $FieldUsedChecked1 />
 EOF
+        }
     }
 
     my $FieldLabelEscaped = $Param{LayoutObject}->Ascii2Html(